% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-wrangling.R
\name{compute_cond_prob}
\alias{compute_cond_prob}
\title{Computes a conditional probability for prob table AC, A-C, -AC, -A-C.}
\usage{
compute_cond_prob(distr_wide, prob)
}
\arguments{
\item{distr_wide}{tibble with columns 'AC', 'A-C', '-AC', '-A-C'}

\item{prob}{str P(x|y) where x,y are one of: A, -A, C, -C}
}
\value{
tibble with column p added containing computed conditional probability
}
\description{
Computes a conditional probability for prob table AC, A-C, -AC, -A-C.
}
