% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modelUtils.R
\name{testModel}
\alias{testModel}
\title{This function fits a model and displays information about this model}
\usage{
testModel(trainset, testset, method)
}
\arguments{
\item{trainset}{}

\item{testset}{}

\item{method}{}
}
\value{
a list
}
\description{
This function fits a model and displays information about this model
}

