% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modelUtils.R
\name{testModel}
\alias{testModel}
\title{Test predictive models}
\usage{
testModel(formula, trainset, testset, outcome, method, ...)
}
\arguments{
\item{formula}{The formula used to train the model}

\item{trainset}{The dataset to trin the data with}

\item{testset}{The dataset with test data}

\item{outcome}{The outcome column name}

\item{method}{The training algorithm to use (eg. glm, rpart, rf, gbm...)}

\item{weights}{Define weights as for the train method}

\item{classProbs}{a logical; should class probabilities be computed for classification models (along with predicted values) in each resample?}
}
\value{
a list
}
\description{
This function fits a model and displays information about this model
}

